Support hyperlinks for manually added PDF pages
Old mechanism of putting hyperlinks to various pages based on
link coordinates and first page height didn't work for the case when
pages are added to PDF document programmatically by a client
app. Allow to pass a new format of `enableLinks` which tells the
hyperlinks plugin in which container should the links be searched
in and to which page they should be put.